change size_t types to unsigned

+++ include/sealloc/chunk.h
@@ -28,13 +28,13 @@ typedef struct run_state run_t;
 
 typedef struct chunk_state {
   ll_entry_t entry;
-  size_t free_mem;
+  unsigned free_mem;
   // Stores (reg_size / 16) for small and medium size class
   uint8_t reg_size_small_medium[CHUNK_NO_NODES_LAST_LAYER];
   uint8_t buddy_tree[CHUNK_BUDDY_TREE_SIZE_BYTES];
 } chunk_t;
 
-void *chunk_allocate_run(chunk_t *chunk, size_t run_size, size_t reg_size);
+void *chunk_allocate_run(chunk_t *chunk, unsigned run_size, unsigned reg_size);
 void *chunk_deallocate_run(chunk_t *chunk, run_t *run);
 bool chunk_is_empty(chunk_t *chunk);
 bool chunk_is_full(chunk_t *chunk);
@@ -42,6 +42,6 @@ void chunk_init(chunk_t *chunk, void *heap);
 // TODO: improve buddy tree so that we can implement that efficiently
 /* bool chunk_can_allocate_run(chunk_t *chunk, size_t run_size); */
 void chunk_get_run_ptr(chunk_t *chunk, void *ptr, void **run_ptr,
-                       size_t *run_size, size_t *reg_size);
+                       unsigned *run_size, unsigned *reg_size);
 
 #endif /* _CHUNK_H_ */
